计算总路径，用int

+++ src/main/java/yuan/interview/railroad/graph/algorithm/IndividualPath.java
@@ -44,8 +44,8 @@ public class IndividualPath extends Path {
 	}
 	
 	@Override
-	public long getTotalWeight(String dim) {
-		long result = 0;
+	public int getTotalWeight(String dim) {
+		int result = 0;
 		
 		Vertex curr = begin;
